requirements.txt
torch
numpy
pillow
isort
black
